Refactor: user type 파일명 수정
