+++ src/modules/IdCard/IdCard.client.tsx
@@ -3,7 +3,7 @@
 import Image from 'next/image';
 
 import Tag from '@/components/Tag/Tag';
-import { CharacterType } from '@/types/user';
+import { CharacterType } from '@/types/user.type';
 
 type IdCardProps = {
   idCardId: number;
